Make it a module format
To make use of the existing syslog format that lnav supports by default,
and remove the duplication. Now this format will only be used as a
sub-format, to parse the body of the message, which is formatted by
xapi, and the syslog format will act as a "container" for it.