Add main non-module format in addition to module format
It seems that the module format cannot override the log-level detected
by the container format, in this case usually syslog. Therefore a "main"
non-module format is necessary for accurate log level detection in
xensource.log.
The module format is still there to parse XenServer logs in the syslog.
However, the log level will always be accurate, as it will be determined
by the syslog_log container format.